requirements.txt
pre-commit
# Adafruit_DHT deprecated
adafruit-circuitpython-dht
# prometheus_client
paho-mqtt